+++ packages/flutter_gemma/lib/core/api/inference_installation_builder.dart
@@ -1,4 +1,3 @@
-import 'package:flutter_gemma/core/domain/model_source.dart';
 import 'package:flutter_gemma/core/di/service_registry.dart';
 import 'package:flutter_gemma/core/utils/file_name_utils.dart';
 import 'package:flutter_gemma/flutter_gemma.dart';